Добавлен spring security, логирование в файл

+++ core/src/main/java/liga/medical/personservice/core/repository/PersonDataRepository.java
@@ -1,10 +1,9 @@
 package liga.medical.personservice.core.repository;
 
-import liga.medical.personservice.core.model.PersonData;
+import liga.medical.personservice.core.model.entity.PersonData;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import java.math.BigInteger;
 
 @Repository
-public interface PersonDataRepository extends JpaRepository<PersonData, BigInteger> {
+public interface PersonDataRepository extends JpaRepository<PersonData, Long> {
 }
